+++ TakeRecipeEasily.Infrastructure/Contracts/Commands/RecipesIngredients/RecipeIngredientCreateModel.cs
@@ -5,11 +5,11 @@ namespace TakeRecipeEasily.Infrastructure.Contracts.Commands.RecipesIngredients
 {
     public sealed class RecipeIngredientCreateModel
     {
-        public int Quantity { get; }
+        public double Quantity { get; }
         public Guid IngredientId { get; }
         public Unit Unit { get; }
 
-        public RecipeIngredientCreateModel(int quantity, Guid ingredientId, Unit unit)
+        public RecipeIngredientCreateModel(double quantity, Guid ingredientId, Unit unit)
         {
             Quantity = quantity;
             IngredientId = ingredientId;
