+++ TakeRecipeEasily.Infrastructure/Contracts/Commands/RecipesIngredients/RecipeIngredientUpdateModel.cs
@@ -5,11 +5,11 @@ namespace TakeRecipeEasily.Infrastructure.Contracts.Commands.RecipesIngredients
 {
     public sealed class RecipeIngredientUpdateModel
     {
-        public int Quantity { get; }
+        public double Quantity { get; }
         public Guid IngredientId { get; }
         public Unit Unit { get; }
 
-        public RecipeIngredientUpdateModel(int quantity, Guid ingredientId, Unit unit)
+        public RecipeIngredientUpdateModel(double quantity, Guid ingredientId, Unit unit)
         {
             Quantity = quantity;
             IngredientId = ingredientId;
